Refactor Radar
